% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{data.preproc}
\alias{data.preproc}
\title{preprocess the data}
\usage{
data.preproc(data, is.cat = NULL, levels = 5)
}
\arguments{
\item{data}{An arbitrary dataset (For example data.frame or matrix).}

\item{is.cat}{A boolean list specifies which variables are categorical. (default = NULL)}

\item{levels}{An integer number indicates the maximum levels of categorical variables and it is used when "is.cat" in NULL. (default = 5)}
}
\value{
A normalized data.frame object with specified continuous and (or) categorical variables and no missing values.
}
\description{
Specify categorical and continuous variables and impute the missing values.
}
\author{
Elyas Heidari, Vahid Balazadeh
}
