% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preproc.R
\name{data_preproc}
\alias{data_preproc}
\title{preprocess the data}
\usage{
data_preproc(data, is.cat = NULL, levels = 5)
}
\arguments{
\item{data}{An arbitrary dataset (For example data.frame or matrix).}

\item{is.cat}{A boolean list specifies which variables are categorical. (default = NULL)}

\item{levels}{An integer number indicates the maximum levels of categorical variables. It is used when \code{is.cat} in NULL. (default = 5)}
}
\value{
A normalized data.frame object with specified continuous and (or) categorical variables and no missing values.
}
\description{
Specify categorical and continuous variables and impute the missing values.
}
\examples{
## Using levels
data("Nhanes")
df <- data_preproc(Nhanes, levels = 15)

## Using is.cat
require(datasets)
data("mtcars")
l <- logical(11)
l[c(8, 9)] <- TRUE
df <- data_preproc(mtcars, is.cat = l)

}
\author{
Elyas Heidari, Vahid Balazadeh
}
