% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{data_preprocess}
\alias{data_preprocess}
\title{This function corrects NAN and normalizes data.}
\usage{
data_preprocess(data, impute = T, scale = T, is.cont = NULL)
}
\description{
This function corrects NAN and normalizes data.
}
