% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm}
\alias{dgm}
\title{construct and visualize causal (directed) networks}
\usage{
dgm(data, alpha = 1e-05, dtype = "gaussian", community = TRUE,
  betweenness = TRUE, plot = TRUE, ...)
}
\arguments{
\item{data}{A normalized dataframe or matrix with no missing data.}

\item{alpha}{A threshold for partial correlation thresholding dtype (default = 0.05). To be used only when the dtype "threshold" is used.}

\item{dtype}{Either "guassian" for continuous/ordinal discrete data or "discrete" for discrete/nominal data. (default = "gaussian") See details for more description.}

\item{community}{A logical value to show if the node communities should be detected and colored in the returned graph. (default = TRUE)}

\item{betweenness}{A logical value to show if the node betweenness measurements should be computed and returned from the function. (default = TRUE)}

\item{plot}{A logical value to show if the graph should be plotted. (default = TRUE)}

\item{...}{Remaining parameters of skleton function.}
}
\value{
A list in which each element is the details of a specific fitted dtype.
\item{graph}{an igraph object of the graphical model.}
\item{betweenness}{betweenness measurements of each node.}
\item{communities}{a named vector indicating the community of each node.}
\item{network}{a visNetwork plot of the graphical model.}
}
\description{
A wrapper of skeleton function in pcalg package. Estimates the skeleton of a Directed Acyclic Graph (DAG) using PC, FCI, or RFCI algorithms.
}
\details{
There is no specific distribution needed for the data. The parameter dtype will be used for determining the data type to be provided as the input of the function. However, it is highly recommended to use "guassian" data type for both continuous and ordinal discrete data.
}
\references{
D. Colombo and M.H. Maathuis (2014).Order-independent constraint-based causal structure learning. Journal of Machine Learning Research 15 3741-3782.

D. Colombo, M. H. Maathuis, M. Kalisch, T. S. Richardson (2012). Learning high-dimensional directed acyclic graphs with latent and selection variables. Ann. Statist. 40, 294-321.

M. Kalisch and P. Buehlmann (2007). Estimating high-dimensional directed acyclic graphs with the PC-algorithm, JMLR 8 613-636.

P. Spirtes, C. Glymour and R. Scheines (2000). Causation, Prediction, and Search, 2nd edition, MIT Press.
}
\author{
Elyas Heidari
}
