% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{dim.reduce}
\alias{dim.reduce}
\title{Dimensionality reduction and visualization.}
\usage{
\method{dim}{reduce}(X, method = "tsne", verbose = TRUE)
}
\arguments{
\item{X}{a dataset to be reduced in dimension.}

\item{method}{a character string as the name of the method. Available values are "tsne" (the default), "tsne_rep", "tsne_early", "tsne_late", "umap", "umap_mom",  "umap_l_bfgs".}

\item{verbose}{a logical value. If TRUE (the default) the details will be printed.}
}
\value{
a plot of data points in the 2 dimensional space.
}
\description{
Reduce dimentionality with a method in \{ tsne, umap, repeated tsne, Momenton umap, L-BFGS umap, Early Exaggeration tsne, and Late Exaggeration tsne\}.
}
\author{
Elyas Heidari with a refrence smallvis package
}
