% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{dime.reduce}
\alias{dime.reduce}
\title{Dimensionality reduction and visualization.}
\usage{
dime.reduce(data, method = "pca", annot1 = NULL, annot1.name = "annot1",
  annot2 = NULL, annot2.name = "annot2")
}
\arguments{
\item{data}{a dataset to be reduced in dimension.}

\item{method}{a character string as the name of the method. Available values are "pca" (the default), "tsne", "umap".}

\item{annot1}{Defaults to NULL.}

\item{annot1.name}{Defaults to "annot1".}

\item{annot2}{Defaults to NULL.}

\item{annot2.name}{Defaults to "annot2".}
}
\value{
A plot of data points in the 2 dimensional space.
}
\description{
Reduce dimensionality with a method in \{tsne, umap, pca\}.
}
\author{
Elyas Heidari
}
