% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.R
\name{div}
\alias{div}
\title{calculating Kullback-Leibler divergence for each categorical variable between two different groups of samples.}
\usage{
div(data, var, permute = 0)
}
\arguments{
\item{data}{a dataframe of categorical variables.}

\item{var}{a character string or an integer indicating the column of the variable of interest. Note that the column should be of two levels (binary-valued).}

\item{permute}{an integer indicating the number of permutations for permutation test. If 0 (the default) no permutation test will be carried out.}
}
\value{
if permute = 0 returns a dataframe including Kullback-Liebler (KL) divergence. if permute > 0 returns a dataframe including KL divergence and p.values.
}
\description{
A function to calculate Kullback-Leibler divergence for each categorical variable between two groups of samples which have different values of a binary-valued variable.
}
\details{
The function helps users to find out the variables with the most divergence between two groups with different states of one specific variable. For instance, within a dataset of health measurements, we are interested in finding the most important variables in occurring cardiovascular disease.
The function is able to carry out permutation test to calculate the p_value for each variable.
}
\examples{

}
\author{
Elyas Heidari
}
