% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_2.R
\name{div2}
\alias{div2}
\title{calculating Kullback-Leibler divergence for each categorical variable between two different groups of samples.}
\usage{
div2(data, var1, var2, permute = 0, frac = 0.05)
}
\arguments{
\item{data}{a dataframe of categorical variables.}

\item{var1}{a vector of continuous values indicating the first variable.}

\item{var2}{a vector of continuous values indicating the second variable.}

\item{permute}{an integer indicating the number of permutations for permutation test. If 0 (the default) no permutation test will be carried out.}

\item{frac}{a double value between 0 and 1 which indicates the fraction of outliers in the fit model. That is, the threshold to recognize a datapoint as an outlier of the fit line.}
}
\value{
if permute = 0 returns a dataframe including Kullback-Liebler (KL) divergence. if permute > 0 returns a dataframe including KL divergence and p.values.
}
\description{
A function to calculate Kullback-Leibler divergence for each categorical variable between two groups of samples which do not follow the linear association between two continuous variables.
}
\examples{
TODO

}
\author{
Elyas Heidari
}
