% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_repres.R
\name{find_repres}
\alias{find_repres}
\title{Find representative variables.}
\usage{
find_repres(graph, ratio = 0.1, weighted = FALSE)
}
\arguments{
\item{graph}{An arbitrary graph object of data in R.}

\item{ratio}{A number between (0, 1) indicating the ratio of variables to be chosen as representative. Defaults to 0.1.}

\item{weighted}{A logical indicating if the number of nodes selected from each community is weighted. Defaults to FALSE. See details for more information.}
}
\value{
a vector containing names of representative variables.
}
\description{
Uses \code{\link[muvis]{graph_vis}} community detection on the input graph and selects a number of variables from each community as representative.
}
\details{
The function uses \code{\link[muvis]{graph_vis}} to detect communities of variables in the \code{graph} input. It sorts each community nodes based on their degree and selects an equal number of each based on \code{ratio} parameter.
If \code{weighted} is TRUE it will select different number of nodes from each community based on their sizes.
}
\examples{
data("NHANES")
## Using \\code{\\link[muvis]{min_forest}} graph
mf <- min_forest(data = NHANES[1:1000, ], levels = 10)
repres_vars <- find_repres(graph = mf$graph, ratio = .2)

## Using \\code{\\link[muvis]{ggm}} graph with \\code{weighted = TRUE}
gm <- ggm(data = NHANES[1:1000, ], levels = 10)
repres_vars <- find_repres(graph = gm$graph, weighted = TRUE)

}
\author{
Vahid Balazadeh, Elyas Heidari
}
