% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{ggm}
\alias{ggm}
\title{construct and visualize a Gaussian graphical model indicating the associations between continuous variables in data}
\usage{
ggm(data, treshold = 0.05, significance = 0.05, rho = 0.1)
}
\arguments{
\item{data}{A normalized dataframe or matrix with no missing data of continuous measurements.}

\item{significance}{A cutoff for edge significance (default = 0.05).}

\item{rho}{(Non-negative) regularization parameter for glasso (default = 0.1).}

\item{threshold}{A threshold for partial correlation thresholding method (default = 0.05).}
}
\value{
A list in which each element is the details of a specific fitted method.
\item{aic}{Fit details of aic model.}
\item{bic}{Fit details of bic model.}
\item{test}{Fit details of significance test model.}
\item{pcor}{Partial correlation matrix used for thresholding method.}
\item{significance}{Significance matrix consisting p.values of each edge.}
\item{glasso}{Partial correlation matrix calculated by glasso.}
\item{betweenness}{Betweenness measurement for each node.}
}
\description{
A function to fit a Gaussian graphical model to continuous variables using six methods, i.e, stepwise AIC, stepwise BIC, stepwise significance test, partial correlation thresholding, edgewise significance test, and glasso.
The function also visualizes the graphical model and decompose the graph into the underlying communities.
}
\details{
The function combines six methods to construct the model, that is, the edge set is the intersection of all edge sets each of which is found by a method. The package gRim is used to implement AIC, BIC, and stepwise significance test. The method glasso from the package glasso is used to provide a sparse inverse covariance matrix.
}
\references{
Højsgaard, S., Edwards, D., & Lauritzen, S. (2012). Graphical Models with R. Springer US. \url{https://doi.org/10.1007/978-1-4614-2299-0}

Friedman, J., Hastie, T., & Tibshirani, R. (2007). Sparse inverse covariance estimation with the graphical lasso. Biostatistics, 9(3), 432–441. \url{https://doi.org/10.1093/biostatistics/kxm045}

Abreu, G. C. G., Edwards, D., & Labouriau, R. (2010). High-Dimensional Graphical Model Search with thegRapHDRPackage. Journal of Statistical Software, 37(1). \url{https://doi.org/10.18637/jss.v037.i01}
}
\author{
Elyas Heidari
}
