% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_categorical.R
\name{is.categorical}
\alias{is.categorical}
\title{Specify categorical and continuous variables in a dataset.}
\usage{
is.categorical(x, is.cat)
}
\arguments{
\item{x}{An arbitrary dataframe}

\item{is.cat}{A list containing boolean elements related to dataset variables (For each categorical variable it is TRUE)}
}
\value{
A dataset with specified categorical and continuous variables.
}
\description{
Specify categorical and continuous variables in a dataset.
}
\author{
Elyas Heidari
}
