% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm}
\alias{llm}
\title{construct and visualize Log Linear Model.}
\usage{
llm(data, threshold = 0.05, significance = 0.05, methods = c("bic"),
  community = TRUE)
}
\arguments{
\item{data}{A normalized dataframe or matrix with no missing data of continuous measurements.}

\item{threshold}{A threshold for partial correlation thresholding method (default = 0.05). To be used only when the method "threshold" is used.}

\item{significance}{A cutoff for edge significance (default = 0.05). To be used only when the method "significance" is used.}

\item{methods}{(default = "glasso")}

\item{community}{(default = TRUE)}
}
\value{
A list in which each element is the details of a specific fitted method.
\item{graph}{an igraph object of the  model.}
\item{betweenness}{betweenness measurements of each edge.}
\item{network}{a highcharter plot of the  model.}
}
\description{
Fits a Log Linear Model to categorical-valued dataset and visualizes the fitted Graphical Model.
}
\author{
Elyas Heidari
}
