% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minForest.R
\name{min.forest}
\alias{min.forest}
\title{construct and visualize a minimum forest based on Chow-Liu algorithm}
\usage{
min.forest(data, stat = "BIC", community = TRUE, plot = FALSE)
}
\arguments{
\item{data}{a normalized dataframe or matrix with no missing data of continuous and (or) categorical measurements.}

\item{stat}{measure to be minimized: LR, AIC, or BIC (the default). Default is BIC. It can also be a user-defined function with the format: FUN(model, dataset, previous, forbEdges); where the parameters are defined as in chStat. The function must return a structure as in chStat.}

\item{community}{a logical value. If TRUE (the default) the network will be colored into communities of edge-dense subgraphs.}

\item{plot}{(default = FALSE)}
}
\value{
a list containing:
\item{}{a gRapHD object which is the fit model.}
\item{}{betweenness measurements for each node.}
\item{}{the highcharter plot of the network.}
}
\description{
Fits a minimum forest to data.
Also visualizes the the minimum forest.
}
\details{
The function is a wrapper for bnlearn package implementing several algorithms including Constraint-based algorithms (i.e., Max-Min Parents and Children, Semi-Interleaved HITON-PC, and Grow-Shrink), Score-based algorithms (i.e., Hill-Climbing and Tabu Search), and Hybrid algorithms (i.e., Max-Min Hill-Climbing), and Local Discovery algorithms (i.e, Max-Min Parents and Children and ARACNE). If one uses a more than one algorithm, the function combines all of the algorithms and returns a graph based on the combination. The graph is constructed based on the strength of associations calculated by bootstrapping.
}
\references{
Chow, C.K., and Liu, C.N. (1968) Approximating discrete probability distributions with dependence trees. IEEE Transactions on Information Theory, Vol. IT-14, 3:462-7.

Edwards, D., de Abreu, G.C.G. and Labouriau, R. (2010). Selecting high- dimensional mixed graphical models using minimal AIC or BIC forests. BMC Bioinformatics, 11:18.
}
\author{
Elyas Heidari
}
