% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{NHANES}
\alias{NHANES}
\title{The National Health and Nutrition Examination Surveys (NHANES)}
\format{An object of class \code{data.frame} with 4461 rows and 161 columns.}
\source{
\url{https://www.icpsr.umich.edu/icpsrweb/ICPSR/studies/25504/summary}
}
\usage{
NHANES
}
\description{
A program of studies about health and nutrition of people in the United States. The data set here is from NHANES 2005-2006 data set which contains more than 7,000 variables and nearly 10,000 samples. We selected 161 variables including one variable for Id ("SEQN"), 74 continuous and 86 categorical (with two to fifteen levels each) and 4461 samples with the age from 20 to 85 years. The data set has nearly 1 \% missing values.
}
\keyword{datasets}
