% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{interactive visulizaiton tool for datasets.}
\usage{
plot(data, vars, levels = 5, pie = F)
}
\arguments{
\item{data}{a dataframe. It is strongly recommended that the dataframe has no missing data and is preprocessed.}

\item{vars}{a vector of length one or two including the name (or index) of a (tow) column(s) of data.}

\item{levels}{an integer to indicate the maximum levels of categorical variables. The default is 7.}

\item{pie}{a logical variable. If TRUE the function returns a pie chart for one variable.}
}
\value{
if pie = TRUE the function returns pie chart, otherwise, there may be 5 scenarios:
\item{}{one categorical variable: plots the histogram of the variable.}
\item{}{one continuous variable: plots the density plot of the variable.}
\item{}{one categorical and one continuous variable: plots a boxplot of the continuous variable for different levels of the categorical variable.}
\item{}{two continuous variables: plots a scatter plot of two variables.}
}
\description{
A function to interctively visualize one or a pair of the variables.
}
\examples{

}
\author{
Elyas Heidari
}
