% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assoc.R
\name{plot_assoc}
\alias{plot_assoc}
\title{Plot variables.}
\usage{
plot_assoc(data, vars, levels = NULL, interactive = FALSE)
}
\arguments{
\item{data}{A dataframe. It is strongly recommended that the dataframe has no missing data and is preprocessed.}

\item{vars}{A vector of length one or two including the name (or index) of a (row) column(s) of data.}

\item{levels}{An integer value indicating the maximum number of levels of a categorical variable. To be used to distinguish the categorical variable. Defaults to NULL because it is supposed that \code{data} has been preprocessed using \code{\link[muvis]{data_preproc}} and the categorical variables are specified.}

\item{interactive}{Logical indicating if the output should be interactive. Defaults to FALSE.}
}
\value{
There may be 5 scenarios for vars:
\item{One categorical variable}{Plots the barplot of the variable.}
\item{One continuous variable}{Plots the density (histogram) plot of the variable.}
\item{A categorical and a continuous variable}{Plots a boxplot of the continuous variable for different levels of the categorical variable.}
\item{Two continuous variables}{Plots a scatter plot of two variables.}
\item{Two categorical variables}{Plots a relative histogram showing distribution of one variable for each level of the other.}
(Plots interactively If interactive = TRUE).
}
\description{
Plots one or a pair of variables (non) interactively using ggplot2 and highcharter packages.
}
\examples{
## Preprocess the data
\dontshow{
data("Nhanes")
data <- data_preproc(Nhanes, levels = 15)

## Plot (non)interactive for:
## One categorical variable
plot_assoc(data, vars = "PAD600")
plot_assoc(data, vars = "SMD410", interactive = TRUE)

## One continuous variable
plot_assoc(data, vars = "LBXTC")
plot_assoc(data, vars = "BMXBMI", interactive = TRUE)

## One continuous and one categorical variable
plot_assoc(data, vars = c("LBXTC", "RIAGENDR"))
plot_assoc(data, vars = c("BMXBMI", "PAD600"), interactive = TRUE)
##  Two continuous variables
plot_assoc(data, vars = c("LBXTC", "BMXBMI"))
plot_assoc(data, vars = c("LBXVIE", "LBXVIC"), interactive = TRUE)

## Two categorical variables
plot_assoc(data, vars = c("SMD410", "PAD600"))
plot_assoc(data, vars = c("PAD600", "SMD410"), interactive = TRUE)

## With raw data
plot_assoc(Nhanes, vars = "RIDAGEYR", levels = 15)
}

\donttest{
#' data("Nhanes")
data <- data_preproc(Nhanes, levels = 15)

## Plot (non)interactive for:
## One categorical variable
plot_assoc(data, vars = "PAD600")
plot_assoc(data, vars = "SMD410", interactive = TRUE)

## One continuous variable
plot_assoc(data, vars = "LBXTC")
plot_assoc(data, vars = "BMXBMI", interactive = TRUE)

## One continuous and one categorical variable
plot_assoc(data, vars = c("LBXTC", "RIAGENDR"))
plot_assoc(data, vars = c("BMXBMI", "PAD600"), interactive = TRUE)
##  Two continuous variables
plot_assoc(data, vars = c("LBXTC", "BMXBMI"))
plot_assoc(data, vars = c("LBXVIE", "LBXVIC"), interactive = TRUE)

## Two categorical variables
plot_assoc(data, vars = c("SMD410", "PAD600"))
plot_assoc(data, vars = c("PAD600", "SMD410"), interactive = TRUE)

## With raw data
plot_assoc(Nhanes, vars = "RIDAGEYR", levels = 15)
}

}
\author{
Elyas Heidari, Vahid Balazadeh
}
