% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plt.assoc}
\alias{plt.assoc}
\title{Plot variables.}
\usage{
plt.assoc(data, vars, levels = 5, interactive = FALSE)
}
\arguments{
\item{data}{A dataframe. It is strongly recommended that the dataframe has no missing data and is preprocessed.}

\item{vars}{A vector of length one or two including the name (or index) of a (row) column(s) of data.}

\item{levels}{An integer to indicate the maximum levels of categorical variables. The default is 5.}

\item{interactive}{Logical indicating if the output should be interactive. Defaults to FALSE.}
}
\value{
There may be 5 scenarios for vars:
\item{One categorical variable}{Plots the barplot of the variable.}
\item{One continuous variable}{Plots the density (histogram) plot of the variable.}
\item{A categorical and a continuous variable}{Plots a boxplot of the continuous variable for different levels of the categorical variable.}
\item{Two continuous variables}{Plots a scatter plot of two variables.}
\item{Two categorical variables}{Plots a relative histogram showing distribution of one variable for each level of the other.}
(Plots interactively If interactive = TRUE).
}
\description{
Plots one or a pair of variables (non) interactively using ggplot2 and highcharter packages.
}
\author{
Elyas Heidari, Vahid Balazadeh
}
