% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_impute.R
\name{simple.impute}
\alias{simple.impute}
\title{simple missing-data imputation}
\usage{
simple.impute(data)
}
\arguments{
\item{data}{an arbitrary dataset.}
}
\value{
a normalized dataframe with no missing data of continuous and (or) categorical measurements.
}
\description{
Imputes missing-data using mean value for continuous variables and mode value for categoricals.
}
\author{
Vahid Balazadeh
}
