% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{test.assoc}
\alias{test.assoc}
\title{Test for association between each paired variables and construct heatmap using heatmaply package.}
\usage{
test.assoc(data, vars, levels = 5, plot = FALSE)
}
\arguments{
\item{data}{a dataframe. It is strongly recommended that the dataframe has no missing data and is preprocessed.}

\item{vars}{a list including the name (or index) of columns of data.}

\item{levels}{an integer to indicate the maximum levels of categorical variables. The default is 5.}

\item{plot}{Logical indicating if the heatmap should be constructed. Defaults to FALSE.}
}
\value{
If plot = FALSE, returns a matrix containing p.values of tests between each two variables. Otherwise returns A list which contains:
\item{matrix}{A matrix containing p.values of tests between each two variables.}
\item{heatmap}{A plotly object containing heatmap related to matrix.}
}
\description{
Tests for association between each paired variables:
\describe{
\item{Categorical-Categorical}{Using pearson's chi-squared test (chisq.test function from stats package).}
\item{Continuous-Continuous}{Using correlation test (cor.test function from stats package).}
\item{Categorical-Continuous}{Using analysis of variance model (aov function from stats package).}
}
Also adjusts p.values using Benjamini & Hochberg method (p.adjust function from stats package) and constructs heatmap using heatmaply function.
}
\details{
This provides a wrapper to \code{chisq.test}, \code{cor.test}, \code{aov}, \code{p.adjust} from \code{stats} package to test association between variables
And a wrapper to \code{heatmaply} package to construct heatmap.
}
\author{
Elyas Heidari
}
