% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_assoc.R
\name{test_assoc}
\alias{test_assoc}
\title{Test for association between each paired variables and construct heatmap using heatmaply package.}
\usage{
test_assoc(data, vars, levels = NULL, plot = FALSE)
}
\arguments{
\item{data}{a dataframe. It is strongly recommended that the dataframe has no missing data and is preprocessed.}

\item{vars}{a list including the name (or index) of columns of data.}

\item{levels}{An integer value indicating the maximum number of levels of a categorical variable. To be used to distinguish the categorical variable. Defaults to NULL because it is supposed that \code{data} has been preprocessed using \code{\link[muvis]{data_preproc}} and the categorical variables are specified.}

\item{plot}{Logical indicating if the heatmap should be constructed. Defaults to FALSE.}
}
\value{
If plot = FALSE, returns a matrix containing p.values of tests between each two variables. Otherwise returns A list which contains:
\item{matrix}{A matrix containing p.values of tests between each two variables.}
\item{heatmap}{A plotly object containing heatmap related to matrix.}
}
\description{
Tests for association between each paired variables:
\describe{
\item{Categorical-Categorical}{Using pearson's chi-squared test (chisq.test function from stats package).}
\item{Continuous-Continuous}{Using correlation test (cor.test function from stats package).}
\item{Categorical-Continuous}{Using analysis of variance model (aov function from stats package).}
}
Also adjusts p.values using Benjamini & Hochberg method (p.adjust function from stats package) and constructs heatmap using heatmaply function.
}
\details{
This provides a wrapper to \code{chisq.test}, \code{cor.test}, \code{aov}, \code{p.adjust} from \code{stats} package to test association between variables
And a wrapper to \code{heatmaply} package to construct heatmap.
}
\examples{
data("Nhanes")
## Using raw data
df <- Nhanes[sample(nrow(Nhanes), 1000), ]
test_matrix <- test_assoc(data = df, vars = colnames(df), plot = FALSE, levels = 15)

## Using preprocessed data
data <- data_preproc(Nhanes, levels = 15)
data$SEQN <- NULL
## Outputs the heatmap too (plot = TRUE)
test_mat_heatmap <- test_assoc(data = data, vars = colnames(data[, 1:20]), plot = TRUE)

}
\author{
Elyas Heidari
}
