% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pair.R
\name{test_pair}
\alias{test_pair}
\title{Test for association between two paired variables.}
\usage{
test_pair(data, var1, var2, levels = NULL)
}
\arguments{
\item{data}{A dataframe. It is strongly recommended that the dataframe has no missing data and is preprocessed.}

\item{var1}{First variable}

\item{var2}{Second variable}

\item{levels}{An integer value indicating the maximum number of levels of a categorical variable. To be used to distinguish the categorical variable. Defaults to NULL because it is supposed that \code{data} has been preprocessed using \code{\link[muvis]{data_preproc}} and the categorical variables are specified.}
}
\value{
P.value of test between the two variables.
}
\description{
Tests for association between two variables:
\describe{
\item{Categorical-Categorical}{Using pearson's chi-squared test (chisq.test function from stats package).}
\item{Continuous-Continuous}{Using correlation test (cor.test function from stats package).}
\item{Categorical-Continuous}{Using analysis of variance model (aov function from stats package).}
}
}
\details{
This provides a wrapper to \code{chisq.test}, \code{cor.test}, \code{aov}, \code{p.adjust} from \code{stats} package to test association between two variables
}
\author{
Elyas Heidari
}
