% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{data.preproc}
\alias{data.preproc}
\title{data preprocess}
\usage{
data.preproc(data, is.cat = NULL, levels = 10)
}
\arguments{
\item{data}{an arbitrary dataset.}

\item{is.cat}{A list containing boolean elements related to dataset variables. (For each categorical variable it is TRUE) (Default = NA)}

\item{levels}{an integer to indicate the maximum levels of categorical variables. The default is 10. (It will be used when "is.cat" list in NA)}
}
\value{
a normalized dataframe with no missing data of continuous and (or) categorical measurements.
}
\description{
Specify categorical and continuous variables and imputes dataset missing-values.
}
\author{
Elyas Heidari, Vahid Balazadeh
}
