% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm}
\alias{dgm}
\title{construct and visualize causal (directed) networks}
\usage{
dgm(data, alpha = 1e-05, dtype = "gaussian", community = TRUE,
  betweenness = TRUE, plot = FALSE, levels = NULL, ...)
}
\arguments{
\item{data}{A normalized dataframe or matrix with no missing data.}

\item{alpha}{significance level (number in (0,1)) for the individual conditional independence tests. Defaults to 1e-5.}

\item{dtype}{Either "guassian" for continuous/ordinal discrete data or "discrete" for discrete/nominal data. (default = "gaussian") See details for more description.}

\item{community}{A logical value to show if the node communities should be detected and colored in the returned graph. (default = TRUE)}

\item{betweenness}{A logical value to show if the node betweenness measurements should be computed and returned from the function. (default = TRUE)}

\item{plot}{A logical value to show if the graph should be plotted. (default = FLASE)}

\item{levels}{An integer value indicating the maximum number of levels of a categorical variable. To be used to distinguish the categorical variable.
Defaults to NULL because it is supposed that \code{data} has been preprocessed using \code{\link[muvis]{data_preproc}} and the categorical variables are specified.
If it is set, first will run \code{\link[muvis]{data_preproc}} to specify categorical and continuous variables.}

\item{...}{Remaining parameters of skleton function.}
}
\value{
A list in which each element is the details of a specific fitted dtype.
\item{significanse}{A data.frame containing edges with p.values.}
\item{graph}{an igraph object of the graphical model.}
\item{betweenness}{betweenness measurements of each node.}
\item{communities}{a named vector indicating the community of each node.}
\item{network}{a visNetwork plot of the graphical model.}
}
\description{
A wrapper of skeleton function in pcalg package. Estimates the skeleton of a Directed Acyclic Graph (DAG) using PC, FCI, or RFCI algorithms.
}
\details{
There is no specific distribution needed for the data. The parameter dtype will be used for determining the data type to be provided as the input of the function. However, it is highly recommended to use "guassian" data type for both continuous and ordinal discrete data.
}
\examples{
data("Nhanes")
## Using raw data
## Using "gaussian" method for continuous data
gaussian_dgm <- dgm(data = Nhanes[sample(nrow(Nhanes), 100), ], dtype = "gaussian", levels = 15)

## Using "discrete" method for categorical data
discrete_dgm <- dgm(data = Nhanes[sample(nrow(Nhanes), 100), ], dtype = "discrete", levels = 15)

## Using preprocessed data
data <- data_preproc(Nhanes, levels = 15)
data$SEQN <- NULL
\donttest{
prep_gauss_dgm <- dgm(data = data[sample(nrow(data), 100), ])
}
\dontshow{
prep_gauss_dgm <- dgm(data = data[sample(nrow(data), 100), ], plot = TRUE)
}



}
\references{
D. Colombo and M.H. Maathuis (2014).Order-independent constraint-based causal structure learning. Journal of Machine Learning Research 15 3741-3782.

D. Colombo, M. H. Maathuis, M. Kalisch, T. S. Richardson (2012). Learning high-dimensional directed acyclic graphs with latent and selection variables. Ann. Statist. 40, 294-321.

M. Kalisch and P. Buehlmann (2007). Estimating high-dimensional directed acyclic graphs with the PC-algorithm, JMLR 8 613-636.

P. Spirtes, C. Glymour and R. Scheines (2000). Causation, Prediction, and Search, 2nd edition, MIT Press.
}
\author{
Elyas Heidari
}
