% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{dim_reduce}
\alias{dim_reduce}
\title{Dimensionality reduction and visualization.}
\usage{
dim_reduce(X)
}
\arguments{
\item{X}{a dataset to be reduced in dimension.}
}
\value{
a list of seven plots each of which is for a method mentioned in the description.
}
\description{
A function to reduce dimensionality using seven methods, namely, tsne, umap, repeated tsne, Momenton umap, L-BFGS umap, Early Exaggeration tsne, and Late Exaggeration tsne.
}
\examples{

}
\author{
Elyas Heidari
}
