% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{ggm}
\alias{ggm}
\title{construct and visualize Gaussian Graphical Models.}
\usage{
ggm(data, threshold = 0.05, significance = 0.05, methods = c("glasso"),
  rho = 0.1, community = TRUE)
}
\arguments{
\item{data}{A normalized dataframe or matrix with no missing data of continuous measurements.}

\item{threshold}{A threshold for partial correlation tresholding method (default = 0.05). To be used only when the method "threshold" is used.}

\item{significance}{A cutoff for edge significance (default = 0.05). To be used only when the method "significance" is used.}

\item{rho}{(Non-negative) regularization parameter for glasso (default = 0.1). To be used only when the method "glasso" is used.}

\item{method}{}
}
\value{
A list in which each element is the details of a specific fitted method.
\item{graph}{an igraph object of the graphical model.}
\item{betweenness}{betweenness measurements of each edge.}
\item{network}{a highcharter plot of the graphical model.}
}
\description{
Fit a Gaussian Graphical Model to continuous-valued dataset employing a subset of methods from stepwise AIC, stepwise BIC, stepwise significance test, partial correlation thresholding, edgewise significance test, or glasso.
Also visualizes the fitted Graphical Model.
}
\details{
The function combines the methods to construct the model, that is, the edge set is the intersection of all edge sets each of which is found by a method. The package gRim is used to implement AIC, BIC, and stepwise significance test. The method glasso from the package glasso is used to provide a sparse estimation of the inverse covariance matrix.
}
\references{
Højsgaard, S., Edwards, D., & Lauritzen, S. (2012). Graphical Models with R. Springer US. \url{https://doi.org/10.1007/978-1-4614-2299-0}

Friedman, J., Hastie, T., & Tibshirani, R. (2007). Sparse inverse covariance estimation with the graphical lasso. Biostatistics, 9(3), 432–441. \url{https://doi.org/10.1093/biostatistics/kxm045}

Abreu, G. C. G., Edwards, D., & Labouriau, R. (2010). High-Dimensional Graphical Model Search with thegRapHDRPackage. Journal of Statistical Software, 37(1). \url{https://doi.org/10.18637/jss.v037.i01}
}
\author{
Elyas Heidari
}
