% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_vis.R
\name{graph.vis}
\alias{graph.vis}
\title{graph visualization and community detection}
\usage{
graph.vis(graph, directed = F, community = T, betweenness = T,
  interactive = T)
}
\arguments{
\item{graph}{an arbitrary graph object in R.}

\item{directed}{TRUE if the graph is directed (default = FALSE)}

\item{community}{(default = TRUE)}

\item{betweenness}{(default = TRUE)}

\item{interactive}{(default = TRUE)}
}
\value{
(If interactive = FALSE it plots the noninteractive graph) A list contains:
\item{graph}{an igraph object}
\item{betweenness}{betweenness measurements of each edge. (if betweenness = TRUE)}
\item{network}{a highcharter plot of the graph. (if interactive = TRUE)}
}
\description{
Converts the graph to igraph, finds communities and plots (non)interactivly.
}
\author{
Elyas Heidari, Vahid Balazadeh
}
