% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_2.R
\name{violating.vars}
\alias{violating.vars}
\title{calculating Kullback-Leibler divergence for each categorical variable between two different groups of samples}
\usage{
violating.vars(data, var1, var2, permute = 0, frac = 0.05, levels = 5)
}
\arguments{
\item{data}{A numeric dataframe including no missing value}

\item{var1}{A vector of continuous values indicating the first variable. (the order of values should be the same as the order of rows in data)}

\item{var2}{A vector of continuous values indicating the second variable. (the order of values should be the same as the order of rows in data)}

\item{permute}{An integer indicating the number of permutations for permutation test. If 0 (the default) no permutation test will be carried out.}

\item{frac}{A double value between 0 and 1 which indicates the fraction of outliers in the fit model. That is, the threshold to recognize a data point as an outlier of the fit line.}

\item{levels}{An integer value indicating the maximum number of levels of a categorical variable. To be used to distinguish the categorical variable.}
}
\value{
If permute = 0 returns a dataframe including sorted Kullback-Liebler (KL) divergence. If permute > 0 returns a dataframe including p.values and sorted KL divergence.
}
\description{
Calculates variable-wise Kullback-Leibler divergence between the two groups of samples which violate the linear relationship between two continuous variables.
}
\author{
Elyas Heidari
}
